+++ src/apps/curve/base/curve.factory-stable-pool.token-fetcher.ts
@@ -6,6 +6,4 @@ import { CurveFactoryStablePoolTokenFetcher } from '../common/curve.factory-stab
 export class BaseCurveFactoryStablePoolTokenFetcher extends CurveFactoryStablePoolTokenFetcher {
   groupLabel = 'Pools';
   registryAddress = '0x3093f9b57a428f3eb6285a589cb35bea6e78c336';
-
-  skipVolume = true;
 }