+++ src/apps/dhedge-v2/contracts/viem.contract-factory.ts
@@ -3,7 +3,7 @@ import { Injectable, Inject } from '@nestjs/common';
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { Network } from '~types/network.interface';
 
-import { DhedgeV2Factory__factory, DhedgeV2Token__factory } from './viem';
+import { DhedgeV2Factory__factory, DhedgeV2Staking__factory, DhedgeV2Token__factory } from './viem';
 
 type ContractOpts = { address: string; network: Network };
 
@@ -13,6 +13,9 @@ export class DhedgeV2ViemContractFactory {
 
   dhedgeV2Factory({ address, network }: ContractOpts) {
     return DhedgeV2Factory__factory.connect(address, this.appToolkit.getViemNetworkProvider(network));
+  }
+  dhedgeV2Staking({ address, network }: ContractOpts) {
+    return DhedgeV2Staking__factory.connect(address, this.appToolkit.getViemNetworkProvider(network));
   }
   dhedgeV2Token({ address, network }: ContractOpts) {
     return DhedgeV2Token__factory.connect(address, this.appToolkit.getViemNetworkProvider(network));